File rearrangement and unit test file added
