+++ tablestore/kv_item.go
@@ -1,0 +1,9 @@
+package tablestore
+
+import (
+
+)
+type KVItem struct {
+    Key      []byte
+    Value    []byte
+}
