+++ tablestore/table_schema.go
@@ -1,0 +1,19 @@
+package tablestore
+
+import (
+
+)
+type FleetDBType uint8
+
+const (
+  INT FleetDBType = iota
+  FLOAT
+  TEXT
+)
+
+type FleetDbColumnSpec struct {
+    colname         string
+    coltype         FleetDBType
+    isPartition     bool
+    isClustering    bool
+}
